+++ src/analyze/src/components/pipelines/LoaderSettingsModal.vue
@@ -104,11 +104,19 @@ export default {
           </div>
         </template>
 
+        <div v-if="loader.signupUrl" class="intro-module">
+          <p>This plugin requires an account. If you don't have one, you can <a :href="loader.signupUrl" target="_blank">sign up here</a>.</p>
+        </div>
+
         <ConnectorSettings
           v-if="!isLoadingConfigSettings"
           fieldClass="is-small"
           :config-settings="configSettings"
         />
+
+        <div v-if="loader.docs" class="footnote-module">
+          <p>Need help finding this information? We got you covered with our <a :href="loader.docs" target="_blank">docs here</a>.</p>
+        </div>
 
         <progress
           v-if="isLoadingConfigSettings && !isInstalling"
@@ -130,4 +138,12 @@ export default {
   </div>
 </template>
 
-<style lang="scss"></style>
+<style lang="scss" scoped>
+.intro-module {
+  margin-bottom: 1rem;
+}
+
+.footnote-module {
+  margin-top: 1rem;
+}
+</style>