+++ src/meltano/api/controllers/plugins.py
@@ -1,5 +1,8 @@
 from meltano.core.compiler.project_compiler import ProjectCompiler
-from meltano.core.plugin_discovery_service import PluginDiscoveryService
+from meltano.core.plugin_discovery_service import (
+    PluginDiscoveryService,
+    PluginNotFoundError,
+)
 from meltano.core.plugin import PluginType
 from meltano.core.project import Project
 from meltano.core.project_add_service import ProjectAddService
@@ -22,8 +25,40 @@ def all():
 
 @pluginsBP.route("/installed", methods=["GET"])
 def installed():
+    """Returns JSON of all installed plugins
+
+    Fuses the discovery.yml data with meltano.yml data
+    """
+
     project = Project.find()
-    return jsonify(project.meltano)
+    config = ConfigService(project)
+    discovery = PluginDiscoveryService(project)
+    meltano_manifest = project.meltano
+    installedPlugins = {}
+    plugins = []
+
+    for plugin in config.plugins():
+        try:
+            definition = discovery.find_plugin(plugin.type, plugin.name)
+        except PluginNotFoundError:
+            definition = {}
+
+        merged_plugin_definition = {**definition.canonical(), **plugin.canonical()}
+        merged_plugin_definition.pop("settings", None)
+        merged_plugin_definition.pop("select", None)
+
+        plugins.append(merged_plugin_definition)
+
+    for type in meltano_manifest["plugins"]:
+        installedPlugins[type] = []
+        for type_plugin in meltano_manifest["plugins"][type]:
+            for complete_plugin in plugins:
+                if type_plugin["name"] == complete_plugin["name"]:
+                    installedPlugins[type].append(complete_plugin)
+
+    meltano_manifest["plugins"] = installedPlugins
+
+    return jsonify(meltano_manifest)
 
 
 @pluginsBP.route("/add", methods=["POST"])